consumo de servicios

+++ src/main/java/com/ceiba/ceibaparking/model/Constantes.java
@@ -12,7 +12,7 @@ public class Constantes {
     public static final int VALORDIAMOTO=4000;
     public static final String 	VALIDACIONLETRA = "A";
     public static final int HORASDELDIA = 24;
-    public static final int FACTORMILISEG2HORAS = 3600000;
+    public static final long FACTORMILISEG2HORAS = 3600000;
     
     private Constantes(){
     	throw new IllegalStateException("Utility class");
